Update README.md all CC253x are not recommended
List CC2538 as not recommended and add a note that Z-Stack 3.0.x firmware is not recommended for the CC2530 and CC2531 in a "production" environment (since they are not powerful enough).

The argument is that while CC2538 based Zigbee coordinator adapters works they are no longer recommended by the Zigbee2MQTT community since they use outdated chips and older firmware.  Availability of CC2538 USB adapters are scarce and I also understand that no zigpy devs are using it as Zigbee coordinator reference hardware so not being well tested, (even though CC2538 should work as CC2530 and CC2531 in theory). Thus might be a good idea for zigpy-znp to also just list CC2538 as "not recommended" in favour of newer CC26x2 and CC13x2 based adapters instead so hopefully less new users buy old hardware as their their first Zigbee coordinator. puddly also posted this comment in https://github.com/zigpy/zigpy-znp/issues/48 about CC2538 around a year ago "*The last Z-Stack update for it was released more than two years ago and it will not support the newer route discovery features that improve network stability because the version of Z-Stack that it's stuck with supposedly has glitches.*"

Other references:

https://www.zigbee2mqtt.io/guide/adapters/#not-recommended

https://github.com/Koenkk/Z-Stack-firmware/blob/master/coordinator/